refactor: remove engine, move execute logic to models
