+++ tests/models/test_job.py
@@ -6,8 +6,10 @@ from textwrap import dedent
 import mock
 import pytest
 from pydantic import ValidationError
+from tests.conftest import job_file
 
-from xetl.models.job import Job
+from xetl.models.job import Job, JobDataDirectoryNotFound
+from xetl.models.task import TaskFailure, UnknownTaskError
 from xetl.models.utils.io import InvalidManifestError, ManifestLoadError, parse_yaml
 
 
@@ -1069,3 +1071,172 @@ def test_resolve_variable_circular_placeholders_raises():
             """
         ).strip()
     )
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=0, autospec=True)
+def test_execute_job_simple_job(task_execute, job_manifest_simple_path):
+    Job.from_file(job_manifest_simple_path).execute()
+
+    comands_and_commands = [
+        f"task: {call.args[0].name}, env: {call.args[1]}, dryrun: {call.args[2]}"
+        for call in task_execute.call_args_list
+    ]
+    assert comands_and_commands == [
+        "task: download, env: {'JOB_VAR': 'job-var-value', 'BASE_URL': 'http://example.com/data', 'THROTTLE': 1000, 'OUTPUT': '/tmp/data'}, dryrun: False",
+    ]
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=0, autospec=True)
+def test_execute_job_multiple_commands(task_execute, job_manifest_multiple_commands_path, tasks_fixtures_path, tmpdir):
+    Job.from_file(job_manifest_multiple_commands_path).execute()
+
+    comands_and_commands = [
+        f"task: {call.args[0].name}, env: {call.args[1]}, dryrun: {call.args[2]}"
+        for call in task_execute.call_args_list
+    ]
+    assert comands_and_commands == [
+        # command: Download-File
+        "task: download, env: {'BASE_URL': 'http://example.com/data', 'THROTTLE': 1000, 'OUTPUT': '/tmp/data'}, dryrun: False",
+        # command: Split_File
+        "task: splitter, env: {'FILES': '/tmp/data', 'OUTPUT': '/tmp/data/splits'}, dryrun: False",
+    ]
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=127)
+def test_execute_job_stops_if_command_fails(
+    task_execute, job_manifest_multiple_commands_path, tasks_fixtures_path, tmpdir
+):
+    with pytest.raises(TaskFailure) as excinfo:
+        Job.from_file(job_manifest_multiple_commands_path).execute()
+
+    assert task_execute.call_count == 1, "Task.execute() should have only been called once"
+    assert excinfo.value.returncode == 127, "The exception should contain the return code of the failed task"
+
+
+@pytest.mark.real_verify_data_dir
+@mock.patch("xetl.models.task.Task.execute", mock.Mock(side_effect=Exception("Task should not have been executed")))
+def test_execute_job_fails_if_data_dir_does_not_exist(job_manifest_unknown_data_path):
+    with pytest.raises(JobDataDirectoryNotFound):
+        Job.from_file(job_manifest_unknown_data_path).execute()
+
+
+@pytest.mark.real_verify_data_dir
+def test_execute_job_doesnt_fail_if_data_dir_does_not_exist_when_dryrun(job_manifest_simple_path):
+    Job.from_file(job_manifest_simple_path).execute(dryrun=True)
+
+
+@mock.patch("xetl.models.task.Task.execute")
+def test_execute_job_without_tasks_path_warns(execute_task, tmpdir, caplog):
+    manifest = dedent(
+        """
+        name: Job without manifests
+        data: /data
+        commands: []
+        """
+    )
+    Job.from_file(job_file(manifest, str(tmpdir))).execute()
+    assert "The property `tasks` is not defined in the job manifest, no tasks will be available" in caplog.messages
+
+
+@mock.patch("xetl.models.task.Task.execute")
+def test_execute_job_no_tasks_found(execute_task, tmpdir, caplog):
+    manifest = dedent(
+        """
+        name: Job without manifests
+        data: /data
+        tasks: /tmp/does-not-exist
+        commands: []
+        """
+    )
+    Job.from_file(job_file(manifest, str(tmpdir))).execute()
+    assert "Could not find any tasks at paths ['/tmp/does-not-exist']" in caplog.messages
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=127)
+def test_execute_job_with_unknown_task(task_execute, job_manifest_simple, tasks_fixtures_path, tmpdir):
+    manifest = job_manifest_simple.replace("task: download", "task: unknown")
+    with pytest.raises(UnknownTaskError) as excinfo:
+        Job.from_file(job_file(manifest, tmpdir)).execute()
+
+    assert str(excinfo.value) == "Unknown task `unknown`, should be one of: ['download', 'parser', 'splitter']"
+    task_execute.assert_not_called()
+
+
+@pytest.mark.parametrize(
+    "commands, expected_executed_commands",
+    [
+        (None, ["Download-File", "Split_File"]),
+        ([], []),
+        ("", []),
+        # strings
+        ("download-file,split-file", ["Download-File", "Split_File"]),
+        ("download-file , split-file", ["Download-File", "Split_File"]),
+        ("download-file", ["Download-File"]),
+        ("split-file", ["Split_File"]),
+        ("SPLIT-FILE", ["Split_File"]),
+        ("SPLIT_FILE ", ["Split_File"]),
+        # lists
+        (["download-file", "split-file"], ["Download-File", "Split_File"]),
+        (["download-file"], ["Download-File"]),
+        (["split-file"], ["Split_File"]),
+    ],
+)
+@mock.patch("xetl.models.task.Task.execute", return_value=0)
+def test_execute_job_filtered_commands(
+    task_execute, commands, expected_executed_commands, job_manifest_multiple_commands_path, caplog
+):
+    caplog.set_level(logging.INFO)
+    Job.from_file(job_manifest_multiple_commands_path).execute(commands=commands)
+
+    # Use the logs to decypher which commands executed.. ¯\_(ツ)_/¯
+    # mocking Command.execute does not give us the actual command instance
+    actual_executed_commands = [
+        re.match("Executing command: ([^ ]*)", msg)[1]  # type: ignore
+        for msg in caplog.messages
+        if msg.startswith("Executing command: ")
+    ]
+
+    assert actual_executed_commands == expected_executed_commands
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=0)
+def test_execute_job_filtered_commands_invalid_type(task_execute, job_manifest_multiple_commands_path, caplog):
+    caplog.set_level(logging.INFO)
+    commands = 123
+
+    with pytest.raises(ValueError) as exc:
+        Job.from_file(job_manifest_multiple_commands_path).execute(commands=commands)  # type: ignore
+
+    assert (
+        str(exc.value)
+        == "The `commands` argument must be a list of strings or a comma-separated string of command names"
+    )
+
+
+@mock.patch("xetl.models.task.Task.execute", return_value=0)
+def test_execute_job_skipped_commands_still_resolve(task_execute, tasks_fixtures_path, tmpdir):
+    job_manifest = dedent(
+        f"""
+        name: Multiple job manifest
+        data: /data
+        tasks: {tasks_fixtures_path}
+        commands:
+          - name: skipped
+            task: download
+            skip: true
+            env:
+              BASE_URL: http://example.com/data1
+              THROTTLE: 1000
+              OUTPUT: /tmp/data1/source
+          - name: references-skipped
+            task: splitter
+            env:
+              FILES: ${{job.data}}/files
+              SOURCE: ${{previous.env.OUTPUT}}
+              OUTPUT: /tmp/data1/splits
+        """
+    )
+    Job.from_file(job_file(job_manifest, tmpdir)).execute()
+    assert task_execute.call_count == 1, "Task.execute() should have only been called once"
+    env_arg = task_execute.call_args[0][0]
+    assert env_arg["SOURCE"] == "/tmp/data1/source"
