--- xetl/__init__.py
@@ -1,7 +1,0 @@
-__version__ = "0.0.1"
-
-from xetl import engine
-
-__all__ = [
-    "engine",
-]
