+++ xetl/models/task_test_case.py
@@ -6,6 +6,13 @@ from xetl.models.utils.run import parse_run_command
 
 
 class TaskTestCase(BaseModel):
+    """
+    Defines test cases used to verify that a custom task is functioning as expected. The test case
+    consists of a set of environment variables and a command to execute. The command is expected to
+    return a zero exit code if the task is functioning as expected. If the command returns a non-zero
+    exit code, the task is considered to have failed the test case.
+    """
+
     env: dict[str, EnvVariableType] = {}
     verify: list[str]
